+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.12.4-dev5
+## 0.12.4-dev6
 
 ### Enhancements
