+++ test_unstructured/partition/pdf_image/test_ocr.py
@@ -18,10 +18,6 @@ from unstructured.partition.pdf_image.ocr import pad_element_bboxes
 from unstructured.partition.utils.constants import (
     Source,
 )
-from unstructured.partition.utils.ocr_models.ocr_interface import (
-    get_elements_from_ocr_regions,
-    merge_text_regions,
-)
 from unstructured.partition.utils.ocr_models.paddle_ocr import OCRAgentPaddle
 from unstructured.partition.utils.ocr_models.tesseract_ocr import (
     OCRAgentTesseract,
@@ -231,35 +227,6 @@ def test_aggregate_ocr_text_by_block():
     assert text == expected
 
 
-def test_merge_text_regions(mock_embedded_text_regions):
-    expected = TextRegion.from_coords(
-        x1=437.83888888888885,
-        y1=317.319341111111,
-        x2=1256.334784222222,
-        y2=406.9837855555556,
-        text="LayoutParser: A Unified Toolkit for Deep Learning Based Document Image",
-    )
-
-    merged_text_region = merge_text_regions(mock_embedded_text_regions)
-    assert merged_text_region == expected
-
-
-def test_get_elements_from_ocr_regions(mock_embedded_text_regions):
-    expected = [
-        LayoutElement.from_coords(
-            x1=437.83888888888885,
-            y1=317.319341111111,
-            x2=1256.334784222222,
-            y2=406.9837855555556,
-            text="LayoutParser: A Unified Toolkit for Deep Learning Based Document Image",
-            type=ElementType.UNCATEGORIZED_TEXT,
-        ),
-    ]
-
-    elements = get_elements_from_ocr_regions(mock_embedded_text_regions)
-    assert elements == expected
-
-
 @pytest.mark.parametrize("zoom", [1, 0.1, 5, -1, 0])
 def test_zoom_image(zoom):
     image = Image.new("RGB", (100, 100))
@@ -277,82 +244,6 @@ def mock_layout(mock_embedded_text_regions):
     return [
         LayoutElement(text=r.text, type=ElementType.UNCATEGORIZED_TEXT, bbox=r.bbox)
         for r in mock_embedded_text_regions
-    ]
-
-
-@pytest.fixture()
-def mock_embedded_text_regions():
-    return [
-        EmbeddedTextRegion.from_coords(
-            x1=453.00277777777774,
-            y1=317.319341111111,
-            x2=711.5338541666665,
-            y2=358.28571222222206,
-            text="LayoutParser:",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=726.4778125,
-            y1=317.319341111111,
-            x2=760.3308594444444,
-            y2=357.1698966666667,
-            text="A",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=775.2748177777777,
-            y1=317.319341111111,
-            x2=917.3579885555555,
-            y2=357.1698966666667,
-            text="Unified",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=932.3019468888888,
-            y1=317.319341111111,
-            x2=1071.8426522222221,
-            y2=357.1698966666667,
-            text="Toolkit",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=1086.7866105555556,
-            y1=317.319341111111,
-            x2=1141.2105142777777,
-            y2=357.1698966666667,
-            text="for",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=1156.154472611111,
-            y1=317.319341111111,
-            x2=1256.334784222222,
-            y2=357.1698966666667,
-            text="Deep",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=437.83888888888885,
-            y1=367.13322999999986,
-            x2=610.0171992222222,
-            y2=406.9837855555556,
-            text="Learning",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=624.9611575555555,
-            y1=367.13322999999986,
-            x2=741.6754646666665,
-            y2=406.9837855555556,
-            text="Based",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=756.619423,
-            y1=367.13322999999986,
-            x2=958.3867708333332,
-            y2=406.9837855555556,
-            text="Document",
-        ),
-        EmbeddedTextRegion.from_coords(
-            x1=973.3307291666665,
-            y1=367.13322999999986,
-            x2=1092.0535042777776,
-            y2=406.9837855555556,
-            text="Image",
-        ),
     ]
