+++ unstructured/partition/pdf.py
@@ -872,7 +872,6 @@ def convert_pdf_to_images(
             yield image
 
 
-@requires_dependencies("unstructured_pytesseract", "unstructured_inference")
 def _partition_pdf_or_image_with_ocr(
     filename: str = "",
     file: Optional[Union[bytes, IO[bytes]]] = None,
