+++ unstructured/partition/pdf_image/ocr.py
@@ -1,6 +1,6 @@
 import os
 import tempfile
-from typing import BinaryIO, Dict, List, Optional, Union, cast
+from typing import TYPE_CHECKING, BinaryIO, Dict, List, Optional, Union, cast
 
 import pdf2image
 
@@ -8,12 +8,6 @@ import pdf2image
 # unstructured.documents.elements.Image
 from PIL import Image as PILImage
 from PIL import ImageSequence
-from unstructured_inference.inference.elements import TextRegion
-from unstructured_inference.inference.layout import DocumentLayout, PageLayout
-from unstructured_inference.inference.layoutelement import (
-    LayoutElement,
-)
-from unstructured_inference.models.tables import UnstructuredTableTransformerModel
 
 from unstructured.documents.elements import ElementType
 from unstructured.logger import logger
@@ -29,8 +23,15 @@ from unstructured.partition.utils.constants import (
 )
 from unstructured.partition.utils.ocr_models.ocr_interface import (
     OCRAgent,
-    get_elements_from_ocr_regions,
 )
+from unstructured.utils import requires_dependencies
+
+if TYPE_CHECKING:
+    from unstructured_inference.inference.elements import TextRegion
+    from unstructured_inference.inference.layout import DocumentLayout, PageLayout
+    from unstructured_inference.inference.layoutelement import LayoutElement
+    from unstructured_inference.models.tables import UnstructuredTableTransformerModel
+
 
 # Force tesseract to be single threaded,
 # otherwise we see major performance problems
@@ -91,6 +92,7 @@ def process_data_with_ocr(
         return merged_layouts
 
 
+@requires_dependencies("unstructured_inference")
 def process_file_with_ocr(
     filename: str,
     out_layout: "DocumentLayout",
@@ -127,6 +129,9 @@ def process_file_with_ocr(
     Returns:
         DocumentLayout: The merged layout information obtained after OCR processing.
     """
+
+    from unstructured_inference.inference.layout import DocumentLayout
+
     merged_page_layouts = []
     try:
         if is_image:
@@ -175,6 +180,7 @@ def process_file_with_ocr(
             raise FileNotFoundError(f'File "{filename}" not found!') from e
 
 
+@requires_dependencies("unstructured_inference")
 def supplement_page_layout_with_ocr(
     page_layout: "PageLayout",
     image: PILImage,
@@ -198,7 +204,7 @@ def supplement_page_layout_with_ocr(
             ocr_languages=ocr_languages,
         )
         page_layout.elements[:] = merge_out_layout_with_ocr_layout(
-            out_layout=cast(List[LayoutElement], page_layout.elements),
+            out_layout=cast(List["LayoutElement"], page_layout.elements),
             ocr_layout=ocr_layout,
         )
     elif ocr_mode == OCRMode.INDIVIDUAL_BLOCKS.value:
@@ -236,7 +242,7 @@ def supplement_page_layout_with_ocr(
             raise RuntimeError("Unable to load table extraction agent.")
 
         page_layout.elements[:] = supplement_element_with_table_extraction(
-            elements=cast(List[LayoutElement], page_layout.elements),
+            elements=cast(List["LayoutElement"], page_layout.elements),
             image=image,
             tables_agent=tables.tables_agent,
             ocr_languages=ocr_languages,
@@ -248,13 +254,13 @@ def supplement_page_layout_with_ocr(
 
 
 def supplement_element_with_table_extraction(
-    elements: List[LayoutElement],
+    elements: List["LayoutElement"],
     image: PILImage,
     tables_agent: "UnstructuredTableTransformerModel",
     ocr_languages: str = "eng",
     ocr_agent: OCRAgent = OCRAgent.get_instance(OCR_AGENT_TESSERACT),
     extracted_regions: Optional[List["TextRegion"]] = None,
-) -> List[LayoutElement]:
+) -> List["LayoutElement"]:
     """Supplement the existing layout with table extraction. Any Table elements
     that are extracted will have a metadata field "text_as_html" where
     the table's text content is rendered into an html string.
@@ -324,10 +330,10 @@ def get_table_tokens(
 
 
 def merge_out_layout_with_ocr_layout(
-    out_layout: List[LayoutElement],
-    ocr_layout: List[TextRegion],
+    out_layout: List["LayoutElement"],
+    ocr_layout: List["TextRegion"],
     supplement_with_ocr_elements: bool = True,
-) -> List[LayoutElement]:
+) -> List["LayoutElement"]:
     """
     Merge the out layout with the OCR-detected text regions on page level.
 
@@ -356,8 +362,8 @@ def merge_out_layout_with_ocr_layout(
 
 
 def aggregate_ocr_text_by_block(
-    ocr_layout: List[TextRegion],
-    region: TextRegion,
+    ocr_layout: List["TextRegion"],
+    region: "TextRegion",
     subregion_threshold: float,
 ) -> Optional[str]:
     """Extracts the text aggregated from the regions of the ocr layout that lie within the given
@@ -376,10 +382,11 @@ def aggregate_ocr_text_by_block(
     return " ".join(extracted_texts) if extracted_texts else ""
 
 
+@requires_dependencies("unstructured_inference")
 def supplement_layout_with_ocr_elements(
-    layout: List[LayoutElement],
-    ocr_layout: List[TextRegion],
-) -> List[LayoutElement]:
+    layout: List["LayoutElement"],
+    ocr_layout: List["TextRegion"],
+) -> List["LayoutElement"]:
     """
     Supplement the existing layout with additional OCR-derived elements.
 
@@ -401,10 +408,15 @@ def supplement_layout_with_ocr_elements(
     Note:
     - The function relies on `is_almost_subregion_of()` method to determine if an OCR region
       is a subregion of an existing layout element.
-    - It also relies on `get_elements_from_ocr_regions()` to convert OCR regions to layout elements.
+    - It also relies on `build_layout_elements_from_ocr_regions()` to convert OCR regions to
+     layout elements.
     - The `SUBREGION_THRESHOLD_FOR_OCR` constant is used to specify the subregion matching
      threshold.
     """
+
+    from unstructured.partition.pdf_image.inference_utils import (
+        build_layout_elements_from_ocr_regions,
+    )
 
     ocr_regions_to_remove = []
     for ocr_region in ocr_layout:
@@ -419,7 +431,7 @@ def supplement_layout_with_ocr_elements(
 
     ocr_regions_to_add = [region for region in ocr_layout if region not in ocr_regions_to_remove]
     if ocr_regions_to_add:
-        ocr_elements_to_add = get_elements_from_ocr_regions(ocr_regions_to_add)
+        ocr_elements_to_add = build_layout_elements_from_ocr_regions(ocr_regions_to_add)
         final_layout = layout + ocr_elements_to_add
     else:
         final_layout = layout
@@ -427,7 +439,7 @@ def supplement_layout_with_ocr_elements(
     return final_layout
 
 
-def get_ocr_agent() -> str:
+def get_ocr_agent() -> OCRAgent:
     ocr_agent_module = env_config.OCR_AGENT
     message = (
         "OCR agent name %s is outdated and will be deprecated in a future release; please use %s "
