+++ unstructured/partition/utils/ocr_models/paddle_ocr.py
@@ -1,11 +1,7 @@
-from typing import List
+from typing import TYPE_CHECKING, List
 
 import numpy as np
 from PIL import Image as PILImage
-from unstructured_inference.inference.elements import TextRegion
-from unstructured_inference.inference.layoutelement import (
-    LayoutElement,
-)
 
 from unstructured.documents.elements import ElementType
 from unstructured.logger import logger
@@ -14,6 +10,11 @@ from unstructured.partition.utils.constants import (
     Source,
 )
 from unstructured.partition.utils.ocr_models.ocr_interface import OCRAgent
+from unstructured.utils import requires_dependencies
+
+if TYPE_CHECKING:
+    from unstructured_inference.inference.elements import TextRegion
+    from unstructured_inference.inference.layoutelement import LayoutElement
 
 
 class OCRAgentPaddle(OCRAgent):
@@ -63,7 +64,7 @@ class OCRAgentPaddle(OCRAgent):
 
     def get_layout_from_image(
         self, image: PILImage, ocr_languages: str = "eng"
-    ) -> List[TextRegion]:
+    ) -> List["TextRegion"]:
         """Get the OCR regions from image as a list of text regions with paddle."""
 
         logger.info("Processing entire page OCR with paddle...")
@@ -76,9 +77,12 @@ class OCRAgentPaddle(OCRAgent):
 
         return ocr_regions
 
+    @requires_dependencies("unstructured_inference")
     def get_layout_elements_from_image(
         self, image: PILImage, ocr_languages: str = "eng"
-    ) -> List[LayoutElement]:
+    ) -> List["LayoutElement"]:
+        from unstructured.partition.pdf_image.inference_utils import build_layout_element
+
         ocr_regions = self.get_layout_from_image(
             image,
             ocr_languages=ocr_languages,
@@ -88,13 +92,17 @@ class OCRAgentPaddle(OCRAgent):
         # terms of grouping because we get ocr_text from `ocr_layout, so the first two grouping
         # and merging steps are not necessary.
         return [
-            LayoutElement(
-                bbox=r.bbox, text=r.text, source=r.source, type=ElementType.UNCATEGORIZED_TEXT
+            build_layout_element(
+                bbox=r.bbox,
+                text=r.text,
+                source=r.source,
+                element_type=ElementType.UNCATEGORIZED_TEXT,
             )
             for r in ocr_regions
         ]
 
-    def parse_data(self, ocr_data: list) -> List[TextRegion]:
+    @requires_dependencies("unstructured_inference")
+    def parse_data(self, ocr_data: list) -> List["TextRegion"]:
         """
         Parse the OCR result data to extract a list of TextRegion objects from
         paddle.
@@ -114,6 +122,9 @@ class OCRAgentPaddle(OCRAgent):
         - An empty string or a None value for the 'text' key in the input
           dictionary will result in its associated bounding box being ignored.
         """
+
+        from unstructured.partition.pdf_image.inference_utils import build_text_region_from_coords
+
         text_regions = []
         for idx in range(len(ocr_data)):
             res = ocr_data[idx]
@@ -130,12 +141,12 @@ class OCRAgentPaddle(OCRAgent):
                     continue
                 cleaned_text = text.strip()
                 if cleaned_text:
-                    text_region = TextRegion.from_coords(
+                    text_region = build_text_region_from_coords(
                         x1,
                         y1,
                         x2,
                         y2,
-                        cleaned_text,
+                        text=cleaned_text,
                         source=Source.OCR_PADDLE,
                     )
                     text_regions.append(text_region)
