+++ unstructured/partition/utils/processing_elements.py
@@ -1,11 +1,13 @@
 from collections import defaultdict
-
-from unstructured_inference.inference.layout import DocumentLayout
+from typing import TYPE_CHECKING
 
 from unstructured.partition.utils.constants import Source
 
+if TYPE_CHECKING:
+    from unstructured_inference.inference.layout import DocumentLayout
 
-def clean_pdfminer_inner_elements(document: DocumentLayout) -> DocumentLayout:
+
+def clean_pdfminer_inner_elements(document: "DocumentLayout") -> "DocumentLayout":
     """Clean pdfminer elements from inside tables and stores them in extra_info dictionary
     with the table id as key"""
     defaultdict(list)
